Add files via upload
Iterates through all the times. The time is converted to an integer, in 24-hour format so it can be easily compared to the times in the flight data, which are also 24-hour format. I create a tuple called fieldsList which contains the corresponding time's temperature, humidity, wind speed, and precipitation.